# swagger/openAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Minesweeper API",
    "description": "API for Minesweeper game",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Register a new user",
        "operationId": "register",
        "parameters": [],
        "responses": {
          "201": {
            "description": "User successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "Auth payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPayload"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AuthPayload"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login user",
        "operationId": "login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              }
            }
          },
          "400": {
            "description": "Invalid auth payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "Auth payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPayload"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AuthPayload"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Logout user",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refreshToken",
            "description": "Refresh token",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Refresh tokens",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refreshToken",
            "description": "Refresh token",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token successfully refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuth"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "MongoDB id"
          },
          "username": {
            "type": "string",
            "example": "username"
          },
          "role": {
            "default": "user",
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2022-07-15T14:48:21.027Z"
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "createdAt"
        ],
        "xml": {
          "name": "User"
        }
      },
...
      "AuthPayload": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "xml": {
          "name": "AuthPayload"
        }
      },
      "SuccessAuth": {
        "type": "object",
        "properties": {
          "tokens": {
            "xml": {
              "name": "tokens"
            },
            "$ref": "#/components/schemas/Tokens"
          },
          "user": {
            "xml": {
              "name": "user"
            },
            "$ref": "#/components/schemas/User"
          }
        },
        "xml": {
          "name": "SuccessAuth"
        }
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "jwt refreshToken"
          },
          "accessToken": {
            "type": "string",
            "example": "jwt accessToken"
          }
        },
        "required": [
          "refreshToken",
          "accessToken"
        ],
        "xml": {
          "name": "Tokens"
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "error message"
          },
          "errors": {
            "type": "array",
            "example": [],
            "items": {}
          }
        },
        "required": [
          "message"
        ],
        "xml": {
          "name": "Error"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}